# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gaia API",
    "description": "Gaia Large Knowledge Model API - manages knowledge graph operations",
    "version": "3.0.0",
    "contact": {
      "name": "Gaia API Support",
      "url": "https://github.com/SiliconEinstein/Gaia/issues"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {"url": "http://localhost:8000", "description": "Local development server"}
  ],
  "paths": {
...
    "/commits": {
      "get": {
        "summary": "List all commits",
        "description": "Returns a list of all commits ordered by creation time",
        "responses": {"200": {"description": "List of commits", "content": {"application/json": {"example": {"commits": [{"commit_id": "abc123", "status": "pending_review", "message": "Add findings", "created_at": "2026-03-01T12:00:00Z"}], "total": 1}}}}},
        "tags": ["commits"]
      },
      "post": {
        "summary": "Submit a new commit",
        "description": "Submit a new commit with graph operations",
        "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"message": {"type": "string"}, "operations": {"type": "array"}}}}}, "required": true},
        "responses": {"200": {"description": "Commit submitted"}},
        "tags": ["commits"]
      }
    },
    "/commits/{commit_id}": {
      "get": {
        "summary": "Get commit details",
        "parameters": [{"name": "commit_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Commit details"}, "404": {"description": "Commit not found"}},
        "tags": ["commits"]
      }
    },
    "/commits/{commit_id}/review": {
      "post": {
        "summary": "Review a commit",
        "parameters": [{"name": "commit_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Review result"}, "404": {"description": "Commit not found"}},
        "tags": ["commits"]
      }
    },
    "/commits/{commit_id}/merge": {
      "post": {
        "summary": "Merge a commit",
        "parameters": [{"name": "commit_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Merge result"}, "404": {"description": "Commit not found"}},
        "tags": ["commits"]
      }
    },
...
    "/jobs/{job_id}": {
      "get": {
        "summary": "Get job status",
        "parameters": [{"name": "job_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Job status"}, "404": {"description": "Job not found"}},
        "tags": ["jobs"]
      },
      "delete": {
        "summary": "Cancel job",
        "parameters": [{"name": "job_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Job cancelled"}, "404": {"description": "Job not found"}},
        "tags": ["jobs"]
      }
    },
    "/jobs/{job_id}/result": {
      "get": {
        "summary": "Get job result",
        "parameters": [{"name": "job_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Job result"}, "404": {"description": "Job not found"}, "409": {"description": "Job not completed"}},
        "tags": ["jobs"]
      }
    }
  },
...
}